Moving dashboard navigation tests to dashboard_test.js
